requirement.txt
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.30.2
torch>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
flask
pypdf
python-dotenv
pinecone>=3.0.0
langchain-pinecone
langchain_openai
langchain_experimental
-e .
